Fix LastEventID mismanagement
Sometime in the past, it looks like there was some connection between
MaxMessageID and LastEventID. This connection no longer exists.

At the moment, last_event_id seems to be a queue-based representation of
how many messages have been sent to the queue. So it's essentially an
auto-incrementing ID. I'm not sure what all MaxMessageID does these
days, so there's likely still some problems with this library, but at
least it can now register a queue and properly read messages from it.