Treat 401's as UnauthorizedError as well.
Zulip servers respond with 401's if the authentication fails (e.g. a
bad token), and 403's if the authenticated user does not have
permission to read something.  403's are thus, unlikely to happen for
this method of consuming the API, and 401's are far more likely.

Include 401 responses in UnauthorizedErrors.